10-feb-25-10-20

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -13,22 +13,22 @@ class CategoryController extends Controller
 {
     function __construct()
     {
-         $this->middleware('permission:category-list|category-create|category-edit|category-delete', ['only' => ['index','store']]);
-         $this->middleware('permission:category-create', ['only' => ['create','store']]);
-         $this->middleware('permission:category-edit', ['only' => ['edit','update']]);
-         $this->middleware('permission:category-delete', ['only' => ['destroy']]);
+        $this->middleware('permission:category-list|category-create|category-edit|category-delete', ['only' => ['index', 'store']]);
+        $this->middleware('permission:category-create', ['only' => ['create', 'store']]);
+        $this->middleware('permission:category-edit', ['only' => ['edit', 'update']]);
+        $this->middleware('permission:category-delete', ['only' => ['destroy']]);
     }
 
     public function index(Request $request)
     {
-        $data = Category::orderBy('id','DESC')->with('category')->get();
+        $data = Category::orderBy('id', 'DESC')->with('category')->get();
         // return $data;
-        return view('backEnd.category.index',compact('data'));
+        return view('backEnd.category.index', compact('data'));
     }
     public function create()
     {
-        $categories = Category::orderBy('id','DESC')->select('id','name')->get();
-        return view('backEnd.category.create',compact('categories'));
+        $categories = Category::orderBy('id', 'DESC')->select('id', 'name')->get();
+        return view('backEnd.category.create', compact('categories'));
     }
     public function store(Request $request)
     {
@@ -36,24 +36,24 @@ class CategoryController extends Controller
             'name' => 'required',
             'status' => 'required',
         ]);
-        // image with intervention 
+        // image with intervention
         $image = $request->file('image');
-        if($image){
-        $name =  time().'-'.$image->getClientOriginalName();
-        $name = preg_replace('"\.(jpg|jpeg|png|webp)$"', '.webp',$name);
-        $name = strtolower(preg_replace('/\s+/', '-', $name));
-        $uploadpath = 'public/uploads/category/';
-        $imageUrl = $uploadpath.$name; 
-        $img=Image::make($image->getRealPath());
-        $img->encode('webp', 90);
-        $width = "";
-        $height = "";
-        $img->height() > $img->width() ? $width=null : $height=null;
-        $img->resize($width, $height, function ($constraint) {
-            $constraint->aspectRatio();
-        });
-        $img->save($imageUrl);
-        }else{
+        if ($image) {
+            $name = time() . '-' . $image->getClientOriginalName();
+            $name = preg_replace('"\.(jpg|jpeg|png|webp)$"', '.webp', $name);
+            $name = strtolower(preg_replace('/\s+/', '-', $name));
+            $uploadpath = 'public/uploads/category/';
+            $imageUrl = $uploadpath . $name;
+            $img = Image::make($image->getRealPath());
+            $img->encode('webp', 90);
+            $width = "";
+            $height = "";
+            $img->height() > $img->width() ? $width = null : $height = null;
+            $img->resize($width, $height, function ($constraint) {
+                $constraint->aspectRatio();
+            });
+            $img->save($imageUrl);
+        } else {
             $imageUrl = null;
         }
 
@@ -61,21 +61,22 @@ class CategoryController extends Controller
         $input['slug'] = strtolower(preg_replace('/\s+/', '-', $request->name));
         $input['slug'] = str_replace('/', '', $input['slug']);
 
-        $input['parent_id'] = $request->parent_id?$request->parent_id:0;
+        $input['parent_id'] = $request->parent_id ? $request->parent_id : 0;
         $input['front_view'] = $request->front_view ? 1 : 0;
+        $input['top_category'] = $request->top_category ? 1 : 0;
         $input['image'] = $imageUrl;
         Category::create($input);
-        Toastr::success('Success','Data insert successfully');
+        Toastr::success('Success', 'Data insert successfully');
         return redirect()->route('categories.index');
     }
-    
+
     public function edit($id)
     {
         $edit_data = Category::find($id);
-        $categories = Category::select('id','name')->get();
-        return view('backEnd.category.edit',compact('edit_data','categories'));
+        $categories = Category::select('id', 'name')->get();
+        return view('backEnd.category.edit', compact('edit_data', 'categories'));
     }
-    
+
     public function update(Request $request)
     {
         $this->validate($request, [
@@ -84,46 +85,47 @@ class CategoryController extends Controller
         $update_data = Category::find($request->id);
         $input = $request->all();
         $image = $request->file('image');
-        if($image){
-            // image with intervention 
-            $name =  time().'-'.$image->getClientOriginalName();
-            $name = preg_replace('"\.(jpg|jpeg|png|webp)$"', '.webp',$name);
+        if ($image) {
+            // image with intervention
+            $name = time() . '-' . $image->getClientOriginalName();
+            $name = preg_replace('"\.(jpg|jpeg|png|webp)$"', '.webp', $name);
             $name = strtolower(preg_replace('/\s+/', '-', $name));
             $uploadpath = 'public/uploads/category/';
-            $imageUrl = $uploadpath.$name; 
-            $img=Image::make($image->getRealPath());
+            $imageUrl = $uploadpath . $name;
+            $img = Image::make($image->getRealPath());
             $img->encode('webp', 90);
             $width = "";
             $height = "";
-            $img->height() > $img->width() ? $width=null : $height=null;
+            $img->height() > $img->width() ? $width = null : $height = null;
             $img->resize($width, $height, function ($constraint) {
                 $constraint->aspectRatio();
             });
             $img->save($imageUrl);
             $input['image'] = $imageUrl;
             File::delete($update_data->image);
-        }else{
+        } else {
             $input['image'] = $update_data->image;
         }
         $input['slug'] = strtolower(preg_replace('/\s+/', '-', $request->name));
         $input['slug'] = str_replace('/', '', $input['slug']);
 
-        $input['parent_id'] = $request->parent_id?$request->parent_id:0;
+        $input['parent_id'] = $request->parent_id ? $request->parent_id : 0;
         $input['front_view'] = $request->front_view ? 1 : 0;
-        $input['status'] = $request->status?1:0;
-        
+        $input['top_category'] = $request->top_category ? 1 : 0;
+        $input['status'] = $request->status ? 1 : 0;
+
         $update_data->update($input);
 
-        Toastr::success('Success','Data update successfully');
+        Toastr::success('Success', 'Data update successfully');
         return redirect()->route('categories.index');
     }
- 
+
     public function inactive(Request $request)
     {
         $inactive = Category::find($request->hidden_id);
         $inactive->status = 0;
         $inactive->save();
-        Toastr::success('Success','Data inactive successfully');
+        Toastr::success('Success', 'Data inactive successfully');
         return redirect()->back();
     }
     public function active(Request $request)
@@ -131,14 +133,20 @@ class CategoryController extends Controller
         $active = Category::find($request->hidden_id);
         $active->status = 1;
         $active->save();
-        Toastr::success('Success','Data active successfully');
+        Toastr::success('Success', 'Data active successfully');
         return redirect()->back();
     }
     public function destroy(Request $request)
     {
         $delete_data = Category::find($request->hidden_id);
-        $delete_data->delete();
-        Toastr::success('Success','Data delete successfully');
+        if ($delete_data) {
+            // Delete related products
+            $delete_data->products()->delete();
+
+            // Now delete the category
+            $delete_data->delete();
+        }
+        Toastr::success('Success', 'Data delete successfully');
         return redirect()->back();
     }
 }
